feat(dbaas): ipfiltering & backups (#42)
* feat(dbaas): allowed cidrs schema

* feat(dbaas): resource CRUD ops

* feat(dbaas): read datasource

* docs(dbaas): cidr whitelist example

* fix(dbaas): minor fixes

* fix(dbaas): ip filtering api changes

* feat(dbaas): datasource

* fix(dbaas): update state after update

* feat(dbaas): add kubernetes identifier

* feat(dbaas): implement new datasources and improve api matching (#56)

* refactor(dbaas): clean unused backup resources

* feat(dbaas): replace backups by backup schedule

* docs(dbaas): backup schedule

* feat(dbaas): data datasources

* docs(dbaas): packs & consts datasources

* fix(dbaas): minor fixs

* fix(dbaas): api changes on backup schedules

* fix(dbaas): allowed cidrs type

* refactor: only use int64 (#57)

refactor(kaas): extract schemas

refactor(dbaas): extract schemas

refactor(domain): extract schemas

refactor: unused scopes packges

fix: minor fixes


fix(dbaas): latest api changes

fix: minor fixs

+++ internal/apis/kaas/mock/gen.go
@@ -7,8 +7,8 @@ import (
 	"math/rand/v2"
 )
 
-func genId() int {
-	return rand.Int()
+func genId() int64 {
+	return rand.Int64()
 }
 
 func genKubeconfig() string {
